update types to work for consecutive evals
